@@ -1,27 +1,15 @@
 package by.it.medved.repositories;
 
 import by.it.medved.entities.Ticket;
-import by.it.medved.entities.User;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
 import java.util.List;
 
-public interface TicketRepository {
+public interface TicketRepository extends JpaRepository<Ticket, Long> {
 
-    Ticket createTicket(Ticket ticket);
+    List<Ticket> findTicketsByUserId(Long userId);
 
-    Ticket getTicketById(Long id);
+    List<Ticket> findTicketsByMovieId(Long movieId);
 
-    List<Ticket> getUserTickets(Long userId);
-
-    List<Ticket> getMovieTickets(Long movieId);
-
-    void updateMovieTicket(Long ticketId, LocalDateTime showDateTime, BigDecimal price);
-
-    boolean buyTicket(Long ticketId, User user);
-
-    boolean returnTicket(Long ticketId);
-
-    void deleteTicket(Long ticketId);
+    List<Ticket> findTicketsByMovieIdAndUserNull(Long movieId);
 }
