@@ -1,25 +1,24 @@
 package by.it.medved.util;
 
-public final class Columns {
+import lombok.experimental.UtilityClass;
 
-    public static final String ID = "ID";
-    public static final String ROLE = "ROLE";
-    public static final String LOGIN = "LOGIN";
-    public static final String PASSWORD = "PASSWORD";
-    public static final String FIRST_NAME = "FIRST_NAME";
-    public static final String EMAIL = "EMAIL";
-    public static final String DATE_BIRTHDAY = "DATE_BIRTHDAY";
-    public static final String DATE_CREATED = "DATE_CREATED";
-    public static final String SALT = "SALT";
-    public static final String MOVIE_ID = "MOVIE_ID";
-    public static final String USER_ID = "USER_ID";
-    public static final String MOVIE_TITLE = "MOVIE_TITLE";
-    public static final String SHOW_DATE_TIME = "SHOW_DATE_TIME";
-    public static final String NUMBER_PLACE = "NUMBER_OF_PLACE";
-    public static final String COUNT_FREE_TICKETS = "COUNT_FREE_TICKETS";
-    public static final String PRICE = "PRICE";
-    public static final String AGE_LIMIT = "AGE_LIMIT";
+@UtilityClass
+public class Columns {
 
-    private Columns() {
-    }
+    public final String ID = "ID";
+    public final String ROLE = "ROLE";
+    public final String LOGIN = "LOGIN";
+    public final String PASSWORD = "PASSWORD";
+    public final String FIRST_NAME = "FIRST_NAME";
+    public final String EMAIL = "EMAIL";
+    public final String DATE_BIRTHDAY = "DATE_BIRTHDAY";
+    public final String DATE_CREATED = "DATE_CREATED";
+    public final String SALT = "SALT";
+    public final String MOVIE_ID = "MOVIE_ID";
+    public final String USER_ID = "USER_ID";
+    public final String MOVIE_TITLE = "MOVIE_TITLE";
+    public final String SHOW_DATE_TIME = "SHOW_DATE_TIME";
+    public final String PLACE_NUMBER = "PLACE_NUMBER";
+    public final String PRICE = "PRICE";
+    public final String AGE_LIMIT = "AGE_LIMIT";
 }
