@@ -1,28 +1,25 @@
 package by.it.medved.util;
 
-public final class Message {
+import lombok.experimental.UtilityClass;
 
-    public static final String LOGIN_BUSY_MESSAGE = " -> занят! Придумайте новый логин.";
-    public static final String LOGIN_NOT_REGISTERED = "Введенный логин не зерегистрирован.";
-    public static final String INCORRECT_PASSWORD = "Введен неверный пароль.";
-    public static final String INVALID_AGE = "Вы не можете купить билет, ограничение %d+";
-    public static final String INVALID_DATE_TO_BUY = "Мероприятие уже прошло!";
-    public static final String INVALID_DATE_TO_RETURN = "Tickets can be returned no later than 10 minutes before the start!";
-    public static final String INVALID_PLACES = "No free places.";
-    public static final String INCORRECT_CURRENT_PASSWORD = "Incorrect current password!";
-    public static final String PASSWORDS_MISMATCH = "Passwords mismatch!";
-    public static final String REGEX_PASSWORD = "Пароль должен содержать от 6 до 20 симвлов из цифр и " +
-            "латинских букв верхнего и нижнего регистра";
-    public static final String REGEX_FIRST_NAME = "Имя должно содержать от 4 до 20 симвлов из руских и " +
-            "латинских букв верхнего и нижнего регистра";
-    public static final String REGEX_EMAIL = "Doesn't match email format";
-    public static final String REGEX_DATE = "Date input format: YYYY-MM-DD";
-    public static final String UPDATE_SUCCESSFUL = "Update successful";
-    public static final String BUY_SUCCESSFULLY = "Ticket buy successfully";
-    public static final String RETURN_SUCCESSFULLY = "Ticket return successfully";
-    public static final String USER_TICKETS_EMPTY = "The user has no buyed tickets";
-    public static final String TICKETS_EMPTY = "You has no buyed tickets";
+@UtilityClass
+public class Message {
 
-    private Message() {
-    }
+    public final String EMPTY = "";
+    public final String LOGIN_BUSY_MESSAGE = "Login '%s' is busy! Invent a new login";
+    public final String INCORRECT_PASSWORD = "Password is incorrect";
+    public final String PASSWORDS_MISMATCH = "Passwords mismatch!";
+    public final String USER_BY_ID_NOT_EXIST = "User with id '%s' does not exist";
+    public final String USER_BY_LOGIN_NOT_EXIST = "User with login '%s' does not exist";
+    public final String MOVIE_BY_ID_NOT_EXIST = "Movie with id '%s' does not exist";
+    public final String TICKET_BY_ID_NOT_EXIST = "Ticket with id '%s' does not exist";
+    public final String LOGIN_REQUIREMENTS = "Login must contain between 4 and 30 characters from upper and " +
+            "lower case Latin letters";
+    public final String PASSWORD_REQUIREMENTS = "Password must contain between 6 and 20 characters from " +
+            "numerals and Latin letters of upper and lower case";
+    public final String FIRST_NAME_REQUIREMENTS = "The name must contain between 4 and 20 characters from " +
+            "Russian and Latin upper and lower case letters";
+    public final String EMAIL_REQUIREMENTS = "Doesn't match the email format";
+    public final String DATE_REQUIREMENTS = "Date value should be in the past";
+    public final String TOKEN = "X-API-KEY=7MRRCKR-GZYMBJN-HT1WDBS-G7C0V5C";
 }
